feat: add notification system with fetching, filtering, and marking functionality
